// YOUCOM.ReserVook/src/app/features/master/taxrate/model/taxrateinfo.model.ts
import { Base } from '../../../../shared/model/baseinfo.model'

export class TaxRateInfo implements Base {
  companyNo: string;
  taxRate: number;
  beginDate: string;
  endDate: string;
  taxrateDivision: string;
  status: string;
  version :number;
  creator:string;
  updator:string;
  cdt:string;
  udt:string;
}